update method delete task

+++ src/Controller/TaskController.php
@@ -99,14 +99,30 @@ class TaskController extends AbstractController
      */
     public function deleteTaskAction(Task $task): RedirectResponse
     {
-        if ($this->getUser()->getId() == $task->getUser()->getId()){
-            $this->em->remove($task);
-            $this->em->flush();
+        $rolesUser = $this->getUser()->getRoles();
 
-            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        if ($this->getUser() !== null) {
+            if ($task->getUser() == null && in_array('ROLE_ADMIN', $rolesUser)){
+                $this->em->remove($task);
+                $this->em->flush();
 
-            return $this->redirectToRoute('task_list');
+                $this->addFlash('success', 'La tâche a bien été supprimée par vous administrateur.');
+
+                return $this->redirectToRoute('task_list');
+            }
+
+            if ($task->getUser() !== null){
+                if ($this->getUser()->getId() == $task->getUser()->getId()){
+                    $this->em->remove($task);
+                    $this->em->flush();
+
+                    $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+                    return $this->redirectToRoute('task_list');
+                }
+            }
         }
+
         $this->addFlash('danger', 'Vous ne pouvez pas supprimer cette tâche.');
         return $this->redirectToRoute('task_list');
